Implement pocket with parent collection view
