+++ .experimenter.yaml
@@ -95,8 +95,8 @@ start-at-home-feature:
       type: string
       description: This property provides a default setting for the startAtHomeFeature
       enum:
-        - after-four-hours
         - disabled
+        - after-four-hours
         - always
 tabTrayFeature:
   description: The tab tray screen that the user goes to when they open the tab tray.
