+++ Client/Frontend/Home/Pocket/FxHomePocketDiscoverMoreCell.swift
@@ -1,0 +1,86 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+
+// MARK: - FxHomePocketDiscoverMoreCell
+/// A cell to be placed at the last position in the Pocket section
+class FxHomePocketDiscoverMoreCell: UICollectionViewCell, ReusableCell {
+
+    // MARK: - UI Elements
+    let itemTitle: UILabel = .build { label in
+        label.adjustsFontForContentSizeCategory = true
+        label.font = DynamicFontHelper.defaultHelper.preferredBoldFont(withTextStyle: .title3,
+                                                                       maxSize: FxHomePocketViewModel.UX.discoverMoreMaxFontSize)
+        label.numberOfLines = 0
+        label.textAlignment = .left
+    }
+
+    // MARK: - Variables
+    var notificationCenter = NotificationCenter.default
+
+    // MARK: - Initializers
+    override init(frame: CGRect) {
+        super.init(frame: .zero)
+
+        applyTheme()
+        setupNotifications(forObserver: self,
+                           observing: [.DisplayThemeChanged])
+        setupLayout()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func prepareForReuse() {
+        super.prepareForReuse()
+        itemTitle.text = nil
+    }
+
+    deinit {
+        notificationCenter.removeObserver(self)
+    }
+
+    // MARK: - Helpers
+
+    private func setupLayout() {
+        contentView.layer.cornerRadius = FxHomeHorizontalCellUX.generalCornerRadius
+        contentView.layer.shadowRadius = FxHomeHorizontalCellUX.stackViewShadowRadius
+        contentView.layer.shadowOffset = CGSize(width: 0, height: FxHomeHorizontalCellUX.stackViewShadowOffset)
+        contentView.layer.shadowColor = UIColor.theme.homePanel.shortcutShadowColor
+        contentView.layer.shadowOpacity = 0.12
+
+        contentView.addSubviews(itemTitle)
+        NSLayoutConstraint.activate([
+            itemTitle.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
+            itemTitle.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
+            itemTitle.centerYAnchor.constraint(equalTo: contentView.centerYAnchor)
+        ])
+    }
+}
+
+// MARK: - Theme
+extension FxHomePocketDiscoverMoreCell: NotificationThemeable {
+    func applyTheme() {
+        if LegacyThemeManager.instance.currentName == .dark {
+            itemTitle.textColor = UIColor.Photon.LightGrey10
+        } else {
+            itemTitle.textColor = UIColor.Photon.DarkGrey90
+        }
+
+        contentView.backgroundColor = UIColor.theme.homePanel.recentlySavedBookmarkCellBackground
+    }
+}
+
+// MARK: - Notifiable
+extension FxHomePocketDiscoverMoreCell: Notifiable {
+    func handleNotifications(_ notification: Notification) {
+        switch notification.name {
+        case .DisplayThemeChanged:
+            applyTheme()
+        default: break
+        }
+    }
+}
